+++ src/mlpack/methods/preprocess/scaling_model_impl.hpp
@@ -39,17 +39,17 @@ inline ScalingModel::ScalingModel(const int minvalue,
 inline ScalingModel::ScalingModel(const ScalingModel& other) :
     scalerType(other.scalerType),
     minmaxscale(other.minmaxscale == NULL ? NULL :
-        new MinMaxScaler(*other.minmaxscale)),
+        new MinMaxScaler<>(*other.minmaxscale)),
     maxabsscale(other.maxabsscale == NULL ? NULL :
-        new MaxAbsScaler(*other.maxabsscale)),
+        new MaxAbsScaler<>(*other.maxabsscale)),
     meanscale(other.meanscale == NULL ? NULL :
-        new MeanNormalization(*other.meanscale)),
+        new MeanNormalization<>(*other.meanscale)),
     standardscale(other.standardscale == NULL ? NULL :
-        new StandardScaler(*other.standardscale)),
+        new StandardScaler<>(*other.standardscale)),
     pcascale(other.pcascale == NULL ? NULL :
-        new PCAWhitening(*other.pcascale)),
+        new PCAWhitening<>(*other.pcascale)),
     zcascale(other.zcascale == NULL ? NULL :
-        new ZCAWhitening(*other.zcascale)),
+        new ZCAWhitening<>(*other.zcascale)),
     minValue(other.minValue),
     maxValue(other.maxValue),
     epsilon(other.epsilon)
@@ -93,27 +93,27 @@ inline ScalingModel& ScalingModel::operator=(const ScalingModel& other)
 
   delete minmaxscale;
   minmaxscale = (other.minmaxscale == NULL) ? NULL :
-      new MinMaxScaler(*other.minmaxscale);
+      new MinMaxScaler<>(*other.minmaxscale);
 
   delete maxabsscale;
   maxabsscale = (other.maxabsscale == NULL) ? NULL :
-      new MaxAbsScaler(*other.maxabsscale);
+      new MaxAbsScaler<>(*other.maxabsscale);
 
   delete standardscale;
   standardscale = (other.standardscale == NULL) ? NULL :
-      new StandardScaler(*other.standardscale);
+      new StandardScaler<>(*other.standardscale);
 
   delete meanscale;
   meanscale = (other.meanscale == NULL) ? NULL :
-      new MeanNormalization(*other.meanscale);
+      new MeanNormalization<>(*other.meanscale);
 
   delete pcascale;
   pcascale = (other.pcascale == NULL) ? NULL :
-      new PCAWhitening(*other.pcascale);
+      new PCAWhitening<>(*other.pcascale);
 
   delete zcascale;
   zcascale = (other.zcascale == NULL) ? NULL :
-      new ZCAWhitening(*other.zcascale);
+      new ZCAWhitening<>(*other.zcascale);
 
   minValue = other.minValue;
   maxValue = other.maxValue;
@@ -168,37 +168,37 @@ void ScalingModel::Fit(const MatType& input)
   if (scalerType == ScalerTypes::STANDARD_SCALER)
   {
     delete standardscale;
-    standardscale = new StandardScaler();
+    standardscale = new StandardScaler<>();
     standardscale->Fit(input);
   }
   else if (scalerType == ScalerTypes::MIN_MAX_SCALER)
   {
     delete minmaxscale;
-    minmaxscale = new MinMaxScaler(minValue, maxValue);
+    minmaxscale = new MinMaxScaler<>(minValue, maxValue);
     minmaxscale->Fit(input);
   }
   else if (scalerType == ScalerTypes::MEAN_NORMALIZATION)
   {
     delete meanscale;
-    meanscale = new MeanNormalization();
+    meanscale = new MeanNormalization<>();
     meanscale->Fit(input);
   }
   else if (scalerType == ScalerTypes::MAX_ABS_SCALER)
   {
     delete maxabsscale;
-    maxabsscale = new MaxAbsScaler();
+    maxabsscale = new MaxAbsScaler<>();
     maxabsscale->Fit(input);
   }
   else if (scalerType == ScalerTypes::PCA_WHITENING)
   {
     delete pcascale;
-    pcascale = new PCAWhitening(epsilon);
+    pcascale = new PCAWhitening<>(epsilon);
     pcascale->Fit(input);
   }
   else if (scalerType == ScalerTypes::ZCA_WHITENING)
   {
     delete zcascale;
-    zcascale = new ZCAWhitening(epsilon);
+    zcascale = new ZCAWhitening<>(epsilon);
     zcascale->Fit(input);
   }
 }